fix: fixes after code review
