+++ src/main/java/com/example/bank/registration/ClientController.java
@@ -3,14 +3,12 @@ package com.example.bank.registration;
 import com.example.bank.registration.jpa.Client;
 import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@Slf4j
 class ClientController {
     private final ClientService clientService;
